embedded-service/power/policy: Refactor state machine code
* Move existing state type into own module
* Use action nomenclature to clarify the indended purpose
